Backup keycloak on each ansible playbook run
This dumps the complete keycloak database into a json file on each
ansible run. It does so by starting wildfly on non-default ports
in an async task with the proper args. The keycloak export job gets
started in a fire-and-forget manner and the process terminates later
when the backup has been made. This should work on both master and
slave envionments since the exportruns as it's own process and does
not rely on a running wildfly server to compelte the export.